cloud_functions/requirements.txt
functions-framework==3.*
google-api-python-client>=1.7.8,<2
google-cloud-aiplatform[pipelines]